+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,28 @@
 
 Release History
 ===============
+
+0.5.5
++++++
+* Add support for Azure KeyVault Secrets Provider as an AKS addon
+
+0.5.4
++++++
+* Add operations of maintenance configuration
+
+0.5.3
++++++
+* Add `--enable-pod-identity-with-kubenet` for enabling AAD Pod Identity in Kubenet cluster
+* Add `--fqdn-subdomain parameter` to create private cluster with custom private dns zone scenario
+
+0.5.2
++++++
+* Add support for node public IP prefix ID '--node-public-ip-prefix-id'
+
+0.5.1
++++++
+* Update to use 2021-02-01 api-version
+
 0.5.0
 +++++
 * Modify addon confcom behavior to only enable SGX device plugin by default.